requirements.txt
# Python 3.11.5
fastapi
uvicorn
psycopg2
asyncpg
pydantic
pydantic-settings
python-jose
bcrypt
cryptography
passlib
pytest
pytest-asyncio
httpx